Update sites.txt
adding the list of public basspistol sites.
Established in 2010 as a pro-active reaction to everything that is 
broken in the music-industry. Basspistol is an endlessly evolving 
union for the Underground Artists of the third millenium 
carrying the values of the Creative Commons. We believe in 
Music. The underground is dead, Long Live the Underground!